# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Device": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"identifier": {
						"type": "string"
					},
					"serial_number": {
						"type": "string"
					},
					"battery_level": {
						"type": "number",
						"format": "double"
					},
					"magnetite_level": {
						"type": "number",
						"format": "double"
					},
					"provisioning_status": {
						"type": "string"
					},
					"firmware_version": {
						"type": "string"
					},
					"last_reading_date": {
						"type": "string",
						"format": "date-time"
					},
					"last_contact_date": {
						"type": "string",
						"format": "date-time"
					},
					"service_due_date": {
						"type": "string",
						"format": "date-time"
					},
					"last_cleaned_date": {
						"type": "string",
						"format": "date-time"
					},
					"installed_date": {
						"type": "string",
						"format": "date-time"
					},
					"notifications": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"id",
					"identifier",
					"serial_number",
					"battery_level",
					"magnetite_level",
					"provisioning_status",
					"firmware_version",
					"last_reading_date",
					"last_contact_date",
					"service_due_date",
					"last_cleaned_date",
					"installed_date",
					"notifications"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Organisation": {
				"description": "Organisation objects allow you to associate actions performed\r\nin the system with the organisation that performed them.\r\nThe Organisation object contains common information across\r\nevery organisations in the system regardless of subscription and role.",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "UNique identifier"
					},
					"name": {
						"type": "string",
						"description": "Orgnaisation name"
					},
					"vat_number": {
						"type": "string",
						"description": "compnay vat number"
					},
					"address_line_1": {
						"type": "string",
						"description": "first line address of the compnay"
					},
					"address_city": {
						"type": "string",
						"description": "company city"
					},
					"address_postcode": {
						"type": "string",
						"description": "*\r\ncompany postcode"
					},
					"status": {
						"type": "string",
						"description": "current status pof company"
					},
					"owner": {
						"type": "string",
						"description": "current owner of the company"
					},
					"is_agent": {
						"type": "boolean",
						"description": "whether they are an agent for another company"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"description": "date created in the system"
					}
				},
				"required": [
					"id",
					"name",
					"vat_number",
					"address_line_1",
					"address_city",
					"address_postcode",
					"status",
					"owner",
					"is_agent",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"authenticated": {
						"type": "boolean",
						"enum": [
							false,
							true
						]
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"email",
					"userName",
					"phoneNumbers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_User.email-or-userName-or-phoneNumbers_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"userName",
					"phoneNumbers"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCreationParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-userName-or-phoneNumbers_"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "Api typescript swagger testing",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/devices/{deviceId}": {
			"get": {
				"operationId": "GetDevice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Device"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/organisations/{organisationId}": {
			"get": {
				"operationId": "GetOrganisation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Organisation"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "organisationId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "userName",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}